backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Caching & Task Queue
redis==5.0.1
celery==5.3.4

# Authentication & Security (TC260 Enhanced)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyjwt==2.8.0

# Rate Limiting & Security Middleware
slowapi==0.1.9
python-multipart==0.0.6

# LLM Integrations
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.25.0

# Security Scanning (Development)
bandit==1.7.5
safety==2.3.5
